pptx-html-bridge/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pptx-html-bridge"
version = "0.1.0"
description = "Convert PPTX files to HTML and vice versa"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.6"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-pptx",
    "lxml",
]

[project.urls]
Homepage = "https://github.com/Liyulingyue/pptx-html-bridge"
Repository = "https://github.com/Liyulingyue/pptx-html-bridge"

[project.scripts]
pptx-to-html = "pptx_html_bridge.converter:main"